requirements.txt
asgiref==3.5.2
attrs==21.4.0
backports.zoneinfo==0.2.1
black==22.6.0
click==8.1.3
Django==4.0.6
django-cors-headers==3.13.0 #
django-filter==22.1 #
djangorestframework==3.13.1 #
importlib-metadata==4.12.0
iniconfig==1.1.1 #  s
Markdown==3.4.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
pluggy==1.0.0
psycopg2-binary==2.9.3 #
py==1.11.0
Pygments==2.12.0
pyparsing==3.0.9
pytest==7.1.2
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
tk==0.1.0
tomli==2.0.1
typing_extensions==4.3.0
zipp==3.8.1
